requirements.txt
# Markboard Backend Dependencies

# Web Framework
Flask==3.0.3
flask-cors==4.0.0

# Database
mysql-connector-python==8.2.0

# Authentication & Security  
bcrypt==4.1.2
PyJWT==2.8.0

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.4
pytest-flask==1.3.0
requests==2.31.0